Possibility to pass arguments to the lambda functions via command line
This feature is needed when we want to pass PU profiles that depend on a particular MC sample to puWeightProducer. Example:

'someLambdaFunction,lambdaFunctionWithArguments(arg1;arg2;arg30,anotherLambdaFunction'

The second lambda is automatically bound w/ the arguments provided in the parentheses (in the same order as specified). In other words, the underlying object which the lambda functions instantiate (whether it be bound or unbound) are created with the arguments provided by the user. If there are no arguments (as is the case for the 1st and the 3rd lambda), the default behavior kicks in.

Tested.