Undo JEC in SD mass recomputation only if it's possible
Unfortunately, it turns out that the AK8LS subjet collection does not store the factor that recovers the raw energy of the subjets as required in SD mass computation. However, it's not clear to me if the subjets that we've created for the custom AK8LS collection have JEC applied or not.
Somewhat related: https://github.com/HEP-KBFI/tth-htt/issues/136